Nutrionist AI with Gemini Pro Vision/requirements.txt
streamlit
google.generativeai
python-dotenv
google-cloud-vision